Don't output to STDOUT for successfully sent digests, by default
Most commonly, digests are sent from cron jobs, where any output to
STDOUT is sent to the administrator.  Printing on every successfully
sent digest will thus lead to the administrator getting sent daily mail
which is essentially content-less.  This might lead them to redirect the
output to /dev/null, which would cause them to be ignorant when sending
messages _failed_.

Restrict output to STDOUT to only when messages failed to send, by
default.  The new --verbose option can be used to restore the old
behavior of printing for every message sent.