Properly display transactions dealing with groups in role groups
The new transaction code assumed that the principals added and removed
from role groups were always users; properly handle the case where they
are groups instead.

RT 4.0 and earlier displayed them as "AdminCc GroupName added"; this
changes the wording slightly to add the word "group" before the group
name, to differentiate it from the cases where a user of that name was
added.