Headers and content may be on different attachments when building digests
The recent HTML templates branch causes $attachment->ContentObj for
default templates to no longer contain the top-level RFC822 headers such
as From, To, and Subject, which breaks assumptions in rt-email-digest.

Store the entire transaction, not just the ContentObj attachment, in the
hash.  This allows us to satisfy both the locations which inspect the
headers (making use of the ->Attachments->First), as well as where the
content is extracted (using ->ContentObj->Content).