Remove broken method delegation from the report entry → report
The original methods LabelCode and FindImplementationCode in
RT::Report::Tickets::Entry delegated to methods in RT::Report::Tickets,
but called the methods on the entry object itself rather than a report
object.  This could lead to unexpected behaviour or errors.

In particular, the FindImplementationCode method (used by
LabelValueCode) takes a value and returns a code ref.  If the value is a
plain string, it looks up a method by that name on, crucially, the
calling object.  These method names come from data structures defined in
RT::Report::Tickets, and are expected to refer to methods in that class
as well, not RT::Report::Tickets::Entry.

It hasn't manifested as a problem yet since all the keys that
RT::Report::Tickets::Entry uses happen to point to coderefs instead of
function names.  Future usage of other keys or value switches to strings
would break some reports with runtime errors.

Switch to storing a weak reference to the report object in each entry.
This gives the entry access to methods and data on the report object
itself, removing the need for a delegation scheme and/or redefinition of
report methods in the entry (e.g.  ColumnInfo and ColumnsLists).

The convenience of not going through ->Report-> is minor compared to the
flexibility of having the report object available.