Revert "no need to load object with id of value 0"
ProcessObjectCustomFieldUpdates() is sometimes passed a loaded Object
which may used when the given id is zero.  For example, this is the
usage seen at user creation in Admin/Users/Modify.html.  The new user
object is first created and then passed to our function, but the CF
query args come from the previous page from before the id was known.
Since ->Load(0) _will_ be called on the already loaded Object, the
record class must explicitly subclass Load to make it a no-op when the
identifier is false.  While this is quirky, the need for such a use case
is solid, and more than just user creation takes advantage of it.  It is
essentially the poor man's version of Jifty::Request::Mapper.

Additional CF tests on master which aren't on 4.0-trunk revealed the
regression.

This reverts commit 163a5ed58573e4e5da421342f66fdbabb1ec2506.