Ensure that rights from lifecycles are loaded during RT->Init
Previous to bab8ef2, rights from lifecycles were loaded when
RT::Lifecycle was use'd or require'd -- which, as it happens, was only
in RT->InitClasses.  This is fortunate, because had it been use'd
anywhere, the Lifecycle configuration would have not yet been loaded,
and thus no rights would have been assigned.

bab8ef2 moves rights registration to cache-fill time, which is generally
only upon access of statuses of an object.  This is to late for @ACL
initialdata insersion, however.  Move this earlier, to during ->Init, so
that lifecycle rights are realized early enough.