Don't force Bookmarks to be unlimited.
This code dates from the original Bookmarks implementation when we just
used '__Bookmarks__' instead of "id = '__Bookmarked__'".

Without this, if you have 20 bookmarks, they all list on your homepage.
After the change, it obeys RowsPerPage.

Bookmarks is a normal search now, so you can always click on the title
to see a 'full' list.

I'm leaving __Bookmarks__ alone because it's the old deprecated code
that we're going to remove in 4.4 anyway.