Do not log "About to think about scrips" for non-Tickets
This debug message is misleading, as it implies that scrips might
possibly be able to fire on a non-Ticket object -- omit it entirely for
non-ticket objects.

There is one other case where it was not immediately followed by "About
to prepare scrips": for outgoing mail transactions.  For such cases,
instead emit an explicit "Skipping scrips" debug message.