Create new reminders in the configured "reminder_on_open" status
Otherwise reminders are created in the lifecycle's "on_create" status.
Completing reminders in this case requires a transition between the
"on_create" status (often "new") and the "reminder_on_resolve" status,
which may not be present outside of the stock configuration.

Additionally, creating reminders in a status other than the open status
means reminders start in a state that they can never return to;
re-opening a completed reminder returns to the open status, not the
create state.

Reminders don't need the complexity of a full lifecycle, and a two-state
system of "open"/"complete" is fine.

This does require that the reminder_on_open status is valid at creation
(i.e. can transition from "").  That requirement is less onerous than
requiring a transition directly from a new state to an inactive state.