Don't store information in hidden elements where it can become stale
Storing the username in a hidden form element to submit, and then doing
the redirect server-side, opens the possibility that the hidden element
can become stale -- namely, when the browser's back button is involved.
If a user was selected from the autocomplete, then the browser backed up
to that page, the browser would helpfully remember the value in the
hidden form element.  This would result in any search (but not
autocomplete) bringing the user to the previously autocompleted user.

Instead, perform the redirect client-side, which saves processing a
custom hidden form.  The UserSearch element cannot be simply filled with
the chosen autocomplete value's username because this might result in a
search with more than one result -- leading to a list of matches, and
not the selected user.  Jumping directly to the target page avoids this
problem.