Limit User/Watcher subfields for grouping to public attributes
At the same time, make a few more useful attributes public under the
assumption that if you're putting it into RT, you want to be able to use
it.  The public accessible check still lets _LocalAccessible overlays
specify public => 0.  We start with a static set of fields rather than
RT::User->ReadableAttributes since most of the user attributes are
~useless for grouping.