Allow any provided links in arguments to carry through to ticket creation
Explictly providing locations for SelfService users to add links during
ticket creation is not what most configurations desire; however, there
is value in being able to forge a link that will result in a ticket that
is so linked.  Pass through any linking arguments on the Create page to
the form, where CreateTicket() in SelfService/Display.html will tend to
them appopriately.