Default chart grouping to Status instead of Queue
Queue may be visually crippling on instances with lots of queues.  A
chart by status is less likely to be cluttered and/or broken.  While
statuses may also be extensive on a system with lots of queues and lots
of different lifecycles, a query on such an instance which returns
tickets in all of the above is less likely.

As the default chart is the one shown for any query when the page is
first visited, it is important that it look and function properly.