Handle plot errors with more aplomb
For example, instead of die()ing and showing a broken image when the
chart size is too small to plot, display an error image with the text:

    Error plotting chart: Horizontal size too small

This is, at least, provides something to the user to complain about.  In
the future, it'll be useful when the user can increase the chart size.