CF key and CF to order by must be passed as two different values
0f6ee87 refactored CF ordering into _OrderByCF; in doing so, the new
_OrderByCF function was written to take a set of ordering information
and a custom field.  However, in cases where the CF could not be
uniquely determined, what it was _passed_ was "$queue.$field".  This was
in turn used internally to JOIN against CustomFields.Name, and found no
results.  Because the JOIN was not a LEFT JOIN, this caused no rows to
be returned.

Like _CustomFieldJoin itself, the CF key must be passed separately from
the customfield name or object.  Undo the slightly overzealous
refactoring to force callers to use their additional information to
separate out different CF keys.