Rebuild record attributes after loading plugins
Plugins may provide _LocalAccessible methods in an attempt to modify
which attriutes are public.  However, as InitClasses() is called prior
to InitPlugins(), the _BuildTableAttributes call is too early to be
affected by plugins' changes.

Add an additional _BuildTableAttributes call after plugins are loaded.
In order to preserve the previous behavior, do not remove the previous
callsite.  Otherwise, code which called InitClasses directly would not
have a working set of classes, and load-time code in plugins would not
be able to use the core classes' SearchBuilder-constructed methods.