Default to the "attachment" disposition when none is specified
Attachments received to RT via email usually have a Content-disposition
header, which we continue to respect.

However, attachments uploaded via the web will not and rely on the
default here in AddAttachment().  Previously we relied on MIME::Entity's
default, but it is not appropriate to assume all files are inline-able.

Indeed, common sense dictates that attachments be treated as
attachments, which is in line with user expectations.  Lenient MUAs
which don't always respect the inline disposition have kept us from
noticing when RT sends mail demanding PDFs, sound files, and other
binaries are inlined.