Avoid attempting to nest <a> tags with links and ShowUser
ShowUser now links to the user, which interacts poorly with being inside
the <a href="<% $href %>"> of the thing we are trying to link to.
Rather than link to [ticket, owner, ticket], move the owner to the end
of the line, where it does not occupy such prime link real-estate.