Submit the full formatted email address for auto-suggested new watchers
This lets the phrase be used as the real name if a new user is created.
Otherwise, RT presents the real name and then appears to lose it as soon
as the suggested watcher is added.  It's bugged me for a while, but I
didn't realize the fix was so simple until today.

A few new tests to ensure we don't break this functionality in the
future.