Rename RT::CustomField->_ForObjectType to RegisterLookupType
As custom fields are extensible by plugins and this method is called
outside of RT::CustomField even in core RT, it was already implicitly
public.  Make it explicitly public and name it to match the
now-canonical name "lookup type".  This is, after all, what the column
is called.  "Register" matches the phrasing we've adopted elsewhere.

Renames and privatizes %FRIENDLY_OBJECT_TYPES as well to match.  Other
code should access lookup type and friendly names via the API.

Deprecates _ForObjectType until 4.4.