Sort the list of lookuptypes such that it displays in consistent order
The dropdown of LookupTypes displayed in the Custom field admin
interface is simply pulled from RT::CustomField->LookupTypes, which was
"keys %FRIENDLY_OBJECT_TYPES" Prior to perl 5.18, this meant that the
ordering changed whenever a new lookuptype was added, which was rare;
post-5.18, it changes on every page reload, which makes it completely
unpredictable as a UI element.  This also effectively provided a
different random default LookupType on every custom field created.

Fix the LookupType to be sorted alphabetically.