rt-email-group-admin: Load users by email address after trying name
All of the examples in the usage use --user email@address, although the
individual subcommand usage does state that they are "unames".  While
many users in RT have a username equal to their email address, this is
non-obvious in the provided examples.  Simply make --user take a name or
an email address of an existing user.