Use Query when passed to create subnav links on Charts.html
Charts.html was using the search saved in the session even
when a Query was passed in. This caused incorrect links when
using multiple tabs, looking at search results in one tab, then
looking at a chart in another tab. The session search from the
other tab would be used when creating the Edit Search, Advanced,
etc. links.