Show the Ticket's Subject when modifying the ticket.
This makes it a lot easier to remember what ticket you're on when using
the Links page, and in general improves the usability of these pages.

Fix up the reminders tests that look at the entire ticket title on the
Modify page.

Fixes Ticket/Update.html to have formatting consistent with the rest of
the pages.