Attempt to avoid filesystem race conditions by skipping chdir
During heavily parallel testing, the following may occur:

     Can't cd to (t/tmp/) api-attribute.t-qrUlYS47: No such file or directory
       at t/99-policy.t line 9.
       at /opt/smoke/perl/lib/5.12.3/File/Find.pm line 881.

This is caused by (in this case) the t/tmp/api-attribute.t-qrUlYS47
directory being found in the readdir() call, but having already been
unlink()'d before chdir() can be called.

File::Find allows skipping the chdir() call by setting no_chdir => 1.
This causes $_ to be equivalent to $File::Find::name, which is
irrelevant to our use case.  Remove the race condition by skipping the
chdir(), which does not affect the contents of @files, or the tests run.