Add ids to recorded upgrade entries to track success/failure
Installations and upgrades can run into issues that cause the
upgrade scripts to die. Add ids to track this history and aid
in future troubleshooting.

full_id: An id for multi-step upgrades from 'make upgrade-database'
The full_id tags each individual upgrade step as part of the larger
upgrade to unambiguously show that individual upgrades were run as
part of a full upgrade.

individual_id: An id to bookend the start and end of an upgrade
step and identify if a step failed to complete.

Adds Data::GUID as a guid generator.