SMIME: Ensure that the keyring path is absolute, and exists
If it does not exist, verification and outgoing encryption (anything
that does not require a private key) is possible; as such, warn but do
not disable.

Because PostLoadChecks for SMIME and GnuPG may run more than once (as
called by the Crypt PostLoadCheck), also ensure that Keyring warning
does not occur more than once.