SMIME: Format status into headers
Unlike GnuPG, which logs the raw output of GnuPG into a header and
parses status events from that later, OpenSSL can take the more
straightforward tact of putting a serialized form of the status events
directly into the header.

Use a simple serialization format that does not depend on newlines being
preserved.