Add Date and DateTime CFs to bulk update
Date conversion is needed only for delete because AddValueForObject
in RT::CustomField calls CanonicalizeValue methods to handle
conversion on add.

For Date and DateTime, pass undef for Default to assure that we
don't incorrectly imply that there will be a default value set
if the box is left empty. This inconsistency has been resolved in
4.4 via 4.4/date-input-keep-value

@values is checked for length, otherwise empty string values
are submitted for empty fields resulting in update messages
noting "Not set changed to Not set".