Provide more specific explanations of decryption/verification failures
There may be multiple failures for incoming email messages; previously,
all that was noted was that they existed.  Iterate through every
operation on every part, and display the resulting message if it was a
failure.  This removes the ->error() logging in CheckBadData because it
duplicates messages that have already been printed earlier.