Work around failure of MIME-Tools to round-trip CRLF multiparts
As documented in https://rt.cpan.org/Public/Bug/Display.html?id=60931 ,
MIME-Tools overzealously trims newlines from CRLF-terminated preambles
in multipart messages.  As `openssl smime` outputs CRLF-terminated
messages, this means that the message that openssl outputs is not the
same as the same message as parsed via MIME::Parser and output via
->as_string.  If signature over a multipart message is contained in an
encrypted part, the newline in the multipart preamble will be modified
during parsing, and thus fail to validate.

Add back the LF which is mistakenly removed by the parser, when parsing
MIME parts output from `openssl smime`.  This change should also be
forward-compatible with a MIME::Parser which does not contain the bug.