Prefix headers with X for RFC822 compliance
As documented in issues #20691, Amazon's smarthost service (SES) is
requiring X prefixes, and while we're technically correct according to
RFC2822 and RFC6648, it's easier to be more liberal in what we generate
and not cause problems for users of these services.

This change does not address other problems noted in that ticket, such
as SES claiming that a message/rfc822 cannot be inlined.