Detection of gpg/openssl should not auto-enable the RT configuration
GnuPG and S/MIME integration is the exception rather than the rule in
most deployments -- but the gpg and openssl binaries they depend on are
ever-present.  Additionally, having enabled these two subsystems
(automatically) without having configured them adds little value or
functionality.

Split each into two parts -- the perl dependencies, which are added
whenever the relevant binaries are installed, and the explicit --enable
or --disable, which set the default value in RT_Config.pm and _also_
override the dependency checking.

This allows casual users to later enable the configuration options in
RT_SiteConfig.pm, whereupon the relevant dependencies will already be
installed (assuming the binary existed at configure-time).  For installs
on systems which demand the functionality at configure-time, --enable
will suffice; and lastly, additional dependencies will not be installed
if the binaries are lacking.