Switch the @ACL upgrade (which uses LoadSystemInternalGroup) to a sub
The LoadSystemInternalGroup which would be used believes that the 4.1.13
upgrade steps have already been run.  Obviously, they have not, leading
to failure to find the Privileged group.  Side-step this and use a
subroutine (which also must use DBIx::SearchBuilder::Record::LoadByCols,
or emit well-meaning warnings) to find the group and add the ACL.