Handle more formats of the Forward template.
The original version added in 3.7.15 said "This is forward...of a
ticket" which I changed in 3.8.6 when we split the Templates into
Forward and Forward Ticket.  However, I failed to account for the
missing newline and didn't go back and match the upgrade script with the
current initialdata.

This now matches on a user who installed or ugpraded from an early 3.8
as well as a user who installed on 4.0.

In either case, this is easily manually fixed, but we should try really
hard to fix it for the user.