Ensure that the Attributes table is big enough to contain the upgrades
Prior to RT 4.0.6, the Content column of the Attributes table was only
capable of holding 64k, being a BLOB or TEXT column (which it was
depends on if it is prior to 3.7.87).  If there are a large number of
upgrades which run prior to the 4.0.6 upgrade step which fixes this, the
base64'd data may overflow, and truncate the upgrade history.

Ensure that the first command run in the upgrade process is the relevant
ALTER TABLE to increase the size of the Content column.  This is only
relevant on mysql (see e249f67), and only if the initial version of the
upgrade is prior to 4.0.6.  If rt-setup-database is run more than once,
the ALTER will be idempotent.