Only set X-RT-Encrypt|Sign if we get an argument.
The web ui has a hidden field and a checkbox.  If you check the
checkbox, we get an arrayref and set X-RT-Encrypt: to 1.  If you don't
check the checkbox, we get 0 and we set X-RT-Encrypt to 0.

If you modify the UI so that users cannot uncheck Encrypt (and use
Always Encrypt in your Queue config) this code would set X-RT-Encrypt: 0
and ignore the Queue default.

It now doesn't set a header, and the mail that is sent out uses
WillSignEncrypt to look at the Queue and see that it should encrypt.