Allow S/MIME enccryption to untrusted users with AcceptUntrustedCAs
gpg is incapable of sending mail to untrusted keys in its keyring.  For
this purpose, RT forbids sending mail to keys which have a trust level
of "unknown."  This limitation can be worked around by setting
'trust-model' to 'always' in GnuPGOptions, which forces all keys in the
keyring to be seen as 'ultimitely trusted' -- with the unfortunate
side-effect in such configurations that actually-trusted keys cannot be
differentiated from keys which are not.

S/MIME does not posess the limitation that untrusted keys cannot be used
for encryption.  In the case of AcceptUntrustedCAs, it is in fact a
somewhat expected workflow that an signed but untrusted S/MIME email is
injected, which may wish to be replied to with encryption.

Rather than implement the same solution as GPG (and suffer the
consequent loss of information as to signing trust), omit the TrustLevel
check if AcceptUntrustedCAs is set.