Add upgrade history only if the database is known to be fully set up
Adding upgrade history entries requires that the tables have been
created and ACL'd, and that the coredata have been inserted.  'init' is
not the only way to be prior to that state -- the create and ACL states
can be run by hand, for instance.

Instead of assuming that the database _is_ sufficiently configured,
assume rather that is it not, and thus default to not logging upgrade
history attributes.  Only in cases where InitSystemObjects has already
been called (upgrades and non-coredata insertions) do we have a
guarantee that the database is sufficiently configured, and thus enable
upgrade history logging from then on.