Enable upgrade history logging for all non-RT database changes
There is one further condition, in addition to those in the previous
commit, under which we can be certain that the RT database is
sufficiently configured; namely, in the case of database changes for
extensions.  While logging for 'insert' actions in extensions is already
covered by the previous logic, 'acl' actions were not.

Explicitly connect and set up system objects when conducting database
changes for extensions; this then allows database upgrade steps to be
logged.