Use custom field groupings on SelfService display page
For backwards compatibility, ungrouped CFs are displayed in the "Basics"
box after Basics and "Basics"-specific CFs.  This also moves the table
boundries to resolve [rt3 #26103], aligning all of the labels.

Note that the SelfService update page is unaffected, and does not
respect groupings.  However, all CFs are correctly displayed once, which
was not the case with the display page prior to this change.