Ensure that menu bgcolor (and text color) is applied to not just aileron
The .aileron limitation was in place such that the text color (which
swaps between black and white based on luminance) did not change the
text color of menus in the web2 style -- as those menus are on the left
and not in the "menu bar".  However, the text color _should_ change in
rudder as well, as the menus are also on the menu bar there.  This led
black-on-dark text if the menu color was set too dark in rudder; on
aileron, it switched to white-on-dark text.

Switch from specifying the precise theme, to instead limiting based on
the sf-shadow class, which is only set for horizonal menus; thus, web2
and ballard, which have vertical menus, are correctly exempted from the
change.