Ensure that changes to the __WARN__ handler in included modules to not leak
60068e5 introduced the Test::NoWarnings module, to ensure that tests are
warnings-clean.  However, that module contains logic which runs as soon
as it is require'd which overrides the __WARN__ handler.  Since the
dependency tester attempts to load every module, this caused all
warnings after Test::NoWarnings was loaded to disappear.

Wrap the 'use ModuleName' in a block which local's the warnings handler,
to prevent such changes from leaking.