Warnings cleanup in rt-test-dependencies
The logic added in 6a32fc5 to not display the full @INC path produces
warnings if the error is already undef due to the module existing, but
being too early a version.  This warning was often hidden due to calling
with --with-developer, which loaded Test::NoWarnings and thus silenced
the warning (see ff15f21).

Protect the regex with an defined-ness check.