Don't print anything to STDOUT during ACL granting
When run under the web installer, output to STDOUT is displayed in the
browser.  This led to "Granting access to ..." being displayed atop the
menu during the database initialization step.

MySQL is the only database type whose ACL tries to display output when
successful. As 'rt-setup-database' provides sufficient updates as to
what is going on, an the specific GRANT being run is mostly irrelevant,
omit the warning entirely.  Switch to using the existing logging
infrastructure for the true warnings.