Fix backcompat detection on Oracle, where NULL != 'string' is false
The 4.1.13 backcomapt file is meant to run for all earlier upgrade steps
where the code assumes that limits on groups can be done on Name.
Before it runs, it checks if there are SystemInternal rows where the
Name is different from the Type, to ensure it does not run more than is
necessary.

Unfortunately, this does not work on Oracle, where Name is NULL on
SystemInternal groups -- and Oracle's treatment of NULL means that the
rows thus do not match, and thus the back-compat step does not run.
Because of this, the upgrade process runs into trouble, for instance in
the 4.1.4 upgrade step:

    [error]: Can't load role group AdminCc: Couldn't find row
    [error]: Can't load role group Cc: Couldn't find row
    [error]: Can't load role group Owner: Couldn't find row
    [error]: Can't load role group Requestor: Couldn't find row

Resolve this by alternately checking if the Name of the SystemInternal
groups IS NULL; this is also still a correct alternate condition to
check on other, non-Oracle, databases.