Don't create an empty part on text/plain ticket creations.
We tell ->build the Type of "multipart/mixed" but passing in Content-
Type means that MIME::Entity will make us an empty part of Content-Type:
text/plain. We then attach the *real* text/plain part and then when we
call make_singlepart it returns "can't do that, more than 1 part" and
leaves us with a horrible message.

When this commit was backported in af919efbe it correctly didn't set
Content-Type.

Sample of MIME generated available in issues #27229.