Ensure that Date and DateTime CF values validate using current values
Date and DateTime CFs, for historical reasons, do not place their
defaults in the CF entry box.  Thus, and update where no value is
submitted for such a CF is equivalent to leaving that CF unchanged.
While the CF update code (see RT::Interface::Web line 3088) implements
this, the vlaidation code was not aware.  Thus, a "mandatory" date CF
would not pass validation if left unchanged.

In RT 4.0, this led to silent errors on the Modify page, and did not
prevent other updates.  In RT 4.2, thanks to 16ddc81, failing to enter
the current value explicitly into Date and DateTime CFs causes them to
fail to validate, and thus prevents all other updates of the page.

Alter the validation code to be aware that Date and DateTime CFs, if
empty, should take their values from the current CF values.