Resolve a runtime error if an invalid Owner was passed
In the event that AutocompleteOwners is enabled, users may enter
arbitrary values for Owner.  If the load by id, name, and emailaddress
fails, the value of ->PrincipalObj was undef, leading to runtime errors
when ->id was called on it just below.  This could also happen if an
unloaded user was passed in, which would similarly have an undef
->PrincipalObj.

Guard against this by checking first that the role's value is true,
before checking for its id.